+++ mintapi/api.py
@@ -1,7 +1,4 @@
 from datetime import date, datetime, timedelta
-import email
-import email.header
-import imaplib
 import io
 import json
 import logging
@@ -9,28 +6,14 @@ import os
 import random
 import re
 import requests
-import subprocess
-import sys
 import time
-import zipfile
 import warnings
 
-from selenium.common.exceptions import (
-    ElementNotInteractableException,
-    ElementNotVisibleException,
-    NoSuchElementException,
-    StaleElementReferenceException,
-    TimeoutException,
-)
-from selenium.webdriver import ChromeOptions
-from selenium.webdriver.common.by import By
-from selenium.webdriver.support import expected_conditions
-from selenium.webdriver.support.ui import WebDriverWait
-from seleniumrequests import Chrome
 import xmltodict
-
-import oathtool
 import pandas as pd
+
+from mintapi.signIn import sign_in, _create_web_driver_at_mint_com
+
 
 logger = logging.getLogger("mintapi")
 
@@ -64,6 +47,7 @@ def reverse_credit_amount(row):
     return amount if row["isDebit"] else -amount
 
 
+<<<<<<< HEAD
 def get_email_code(
     imap_account, imap_password, imap_server, imap_folder, debug=False, delete=True
 ):
@@ -602,6 +586,8 @@ def get_web_driver(
     return driver, status_message
 
 
+=======
+>>>>>>> 26f023c61990c6bb8a6b9a70e59da912d45ba452
 IGNORE_FLOAT_REGEX = re.compile(r"[$,%]")
 
 
@@ -863,6 +849,12 @@ class Mint(object):
             headers=self._get_api_key_header(),
         ).json()
 
+    def get_categories(self):
+        return self.get(
+            "{}/pfm/v1/categories".format(MINT_ROOT_URL),
+            headers=self._get_api_key_header(),
+        ).json()["Category"]
+
     def get_accounts(self, get_detail=False):  # {{{
         # Issue service request.
         req_id = self.get_request_id_str()
@@ -975,7 +967,7 @@ class Mint(object):
             # Specifying accountId=0 causes Mint to return investment
             # transactions as well.  Otherwise they are skipped by
             # default.
-            if id > 0 or include_investment:
+            if self._include_investments_with_transactions(id, include_investment):
                 params["accountId"] = id
             if include_investment:
                 params["task"] = "transactions"
@@ -1038,15 +1030,13 @@ class Mint(object):
         # the transaction category ID
         categories = self.get_categories()
         for transaction in result:
-            parent = self.get_category_object_from_id(
+            category = self.get_category_object_from_id(
                 transaction["categoryId"], categories
-            )["parent"]
-            transaction["parentCategoryName"] = (
-                "" if parent["name"] == "Root" else parent["name"]
             )
-            transaction["parentCategoryId"] = (
-                "" if parent["name"] == "Root" else parent["id"]
-            )
+            parent = self._find_parent_from_category(category, categories)
+            transaction["parentCategoryId"] = self.__format_category_id(parent["id"])
+            transaction["parentCategoryName"] = parent["name"]
+
         return result
 
     def get_transactions_csv(
@@ -1065,7 +1055,9 @@ class Mint(object):
         # default.
 
         params = {
-            "accountId": acct if acct > 0 else None,
+            "accountId": acct
+            if self._include_investments_with_transactions(acct, include_investment)
+            else None,
             "startDate": convert_date_to_string(convert_mmddyy_to_datetime(start_date)),
             "endDate": convert_date_to_string(convert_mmddyy_to_datetime(end_date)),
         }
@@ -1168,54 +1160,11 @@ class Mint(object):
 
         return accounts
 
-    def get_categories(self):  # {{{
-        # Get category metadata.
-        req_id = self.get_request_id_str()
-        data = {
-            "input": json.dumps(
-                [
-                    {
-                        "args": {
-                            "excludedCategories": [],
-                            "sortByPrecedence": False,
-                            "categoryTypeFilter": "FREE",
-                        },
-                        "id": req_id,
-                        "service": "MintCategoryService",
-                        "task": "getCategoryTreeDto2",
-                    }
-                ]
-            )
-        }
-        response = self.make_post_request(
-            url=self.build_bundledServiceController_url(),
-            data=data,
-            convert_to_text=True,
-        )
-        if req_id not in response:
-            raise MintException('Could not parse category data: "{}"'.format(response))
-        response = json.loads(response)
-        response = response["response"][req_id]["response"]
-
-        # Build category list
-        categories = {}
-        for category in response["allCategories"]:
-            categories[category["id"]] = category
-
-        return categories
-
-    def get_budgets(self, hist=None):  # {{{
-        # Issue request for budget utilization
-        first_of_this_month = date.today().replace(day=1)
-        eleven_months_ago = (first_of_this_month - timedelta(days=330)).replace(day=1)
-        url = "{}/getBudget.xevent".format(MINT_ROOT_URL)
-        params = {
-            "startDate": convert_date_to_string(eleven_months_ago),
-            "endDate": convert_date_to_string(first_of_this_month),
-            "rnd": Mint.get_rnd(),
-        }
-        response = json.loads(self.get(url, params=params, headers=JSON_HEADER).text)
+    def get_budgets(self, hist=None):
+        response = self.__call_budgets_endpoint()
         categories = self.get_categories()
+        income = response["data"]["income"]
+        spending = response["data"]["spending"]
         if hist is not None:  # version proofing api
 
             def mos_to_yrmo(mos_frm_zero):
@@ -1229,7 +1178,7 @@ class Mint(object):
             elif hist < 1:
                 hist = 1
 
-            bgt_cur_mo = max(map(int, response["data"]["income"].keys()))
+            bgt_cur_mo = max(map(int, income.keys()))
             min_mo_hist = bgt_cur_mo - hist
 
             # Initialize and populate dictionary for return
@@ -1239,63 +1188,74 @@ class Mint(object):
             budgets = {}
             for months in range(bgt_cur_mo, min_mo_hist, -1):
                 budgets[mos_to_yrmo(months)] = {}
-                budgets[mos_to_yrmo(months)]["income"] = response["data"]["income"][
-                    str(months)
-                ]["bu"]
-                budgets[mos_to_yrmo(months)]["spending"] = response["data"]["spending"][
-                    str(months)
-                ]["bu"]
+                budgets[mos_to_yrmo(months)]["income"] = income[str(months)]["bu"]
+                budgets[mos_to_yrmo(months)]["spending"] = spending[str(months)]["bu"]
 
             # Fill in the return structure
             for month in budgets.keys():
                 for direction in budgets[month]:
                     for budget in budgets[month][direction]:
-                        category = self.get_category_object_from_id(
-                            budget["cat"], categories
-                        )
-                        budget["cat"] = category["name"]
-                        budget["parent"] = category["parent"]["name"]
+                        budget = self.__format_budget_categories(budget, categories)
 
         else:
             # Make the skeleton return structure
             budgets = {
-                "income": response["data"]["income"][
-                    str(max(map(int, response["data"]["income"].keys())))
-                ]["bu"],
-                "spend": response["data"]["spending"][
-                    str(max(map(int, response["data"]["spending"].keys())))
-                ]["bu"],
+                "income": income[str(max(map(int, income.keys())))]["bu"],
+                "spend": spending[str(max(map(int, spending.keys())))]["bu"],
             }
 
             # Fill in the return structure
             for direction in budgets.keys():
                 for budget in budgets[direction]:
-                    category = self.get_category_object_from_id(
-                        budget["cat"], categories
-                    )
-                    budget["cat"] = category["name"]
-                    # Uncategorized budget's parent is a string: 'Uncategorized'
-                    if isinstance(category["parent"], dict):
-                        budget["parent"] = category["parent"]["name"]
-                    else:
-                        budget["parent"] = category["parent"]
+                    budget = self.__format_budget_categories(budget, categories)
 
         return budgets
 
+    def __call_budgets_endpoint(self):
+        # Issue request for budget utilization
+        first_of_this_month = date.today().replace(day=1)
+        eleven_months_ago = (first_of_this_month - timedelta(days=330)).replace(day=1)
+        url = "{}/getBudget.xevent".format(MINT_ROOT_URL)
+        params = {
+            "startDate": convert_date_to_string(eleven_months_ago),
+            "endDate": convert_date_to_string(first_of_this_month),
+            "rnd": Mint.get_rnd(),
+        }
+        return json.loads(self.get(url, params=params, headers=JSON_HEADER).text)
+
+    def __format_budget_categories(self, budget, categories):
+        category = self.get_category_object_from_id(budget["cat"], categories)
+        budget["cat"] = category["name"]
+        parent = self._find_parent_from_category(category, categories)
+        budget["parent"] = parent["name"]
+        return budget
+
     def get_category_object_from_id(self, cid, categories):
         if cid == 0:
-            return {"parent": "Uncategorized", "name": "Uncategorized"}
+            return {"parent": "Uncategorized", "depth": 1, "name": "Uncategorized"}
 
-        for i in categories:
-            if categories[i]["id"] == cid:
-                return categories[i]
+        result = filter(
+            lambda category: self.__format_category_id(category["id"]) == str(cid),
+            categories,
+        )
+        category = list(result)
+        return (
+            category[0]
+            if len(category) > 0
+            else {"parent": "Unknown", "depth": 1, "name": "Unknown"}
+        )
 
-            if "children" in categories[i]:
-                for j in categories[i]["children"]:
-                    if categories[i][j]["id"] == cid:
-                        return categories[i][j]
+    def __format_category_id(self, cid):
+        return cid if str(cid).find("_") == "-1" else str(cid)[str(cid).find("_") + 1 :]
 
-        return {"parent": "Unknown", "name": "Unknown"}
+    def _find_parent_from_category(self, category, categories):
+        if category["depth"] == 1:
+            return {"id": "", "name": ""}
+
+        parent = self.get_category_object_from_id(
+            self.__format_category_id(category["parentId"]), categories
+        )
+        return {"id": parent["id"], "name": parent["name"]}
 
     def initiate_account_refresh(self):
         data = {"token": self.token}
@@ -1417,10 +1377,8 @@ class Mint(object):
                 )
         return utilization
 
-
-def get_token(driver: Chrome):
-    value_json = driver.find_element_by_name("javascript-user").get_attribute("value")
-    return json.loads(value_json)["token"]
+    def _include_investments_with_transactions(self, id, include_investment):
+        return id > 0 or include_investment
 
 
 def get_accounts(email, password, get_detail=False):
@@ -1456,7 +1414,7 @@ def initiate_account_refresh(email, password):
 
 if __name__ == "__main__":
     warnings.warn(
-        "Calling command line code from api.py is will be deprecated in a future release.\n"
+        "Calling command line code from api.py will be deprecated in a future release.\n"
         "Please call mintapi directly. For examples, see the README.md",
         DeprecationWarning,
     )