+++ tests/test_endpoints_sign_in.py
@@ -4,6 +4,9 @@ import pytest
 
 import mintapi
 import mintapi.api
+import mintapi.signIn
+
+from tests.test_driver import category_example
 
 
 USERNAME = os.environ.get("MINTAPI_USERNAME", None)
@@ -30,7 +33,7 @@ pytestmark = pytest.mark.skipif(
 @pytest.fixture
 def get_mint_driver() -> mintapi.Mint:
     mint = mintapi.Mint()
-    mint.driver = mintapi.api._create_web_driver_at_mint_com(
+    mint.driver = mintapi.signIn._create_web_driver_at_mint_com(
         HEADLESS,
         SESSION_PATH,
         USE_CHROMEDRIVER_ON_PATH,
@@ -41,7 +44,7 @@ def get_mint_driver() -> mintapi.Mint:
 
 @pytest.fixture
 def do_sign_in(get_mint_driver: mintapi.Mint) -> mintapi.Mint:
-    mintapi.api._sign_in(
+    mintapi.sign_in(
         USERNAME,
         PASSWORD,
         get_mint_driver.driver,
@@ -53,7 +56,7 @@ def do_sign_in(get_mint_driver: mintapi.Mint) -> mintapi.Mint:
 
 
 def test_sign_in(get_mint_driver: mintapi.Mint):
-    mintapi.api._sign_in(
+    mintapi.sign_in(
         USERNAME,
         PASSWORD,
         get_mint_driver.driver,
@@ -70,3 +73,9 @@ def test_investment_endpoint(do_sign_in: mintapi.Mint):
     investment_data = do_sign_in.get_investment_data()[0]
     assert "metaData" not in investment_data
     assert "lastUpdatedDate" in investment_data
+
+
+def test_get_categories(do_sign_in: mintapi.Mint):
+    categories = do_sign_in.get_categories()
+    for key in category_example[0].keys():
+        assert key in categories[0].keys()