+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+1.63
+---
+- fix regression introduced in #384 (#398)
+
+1.62
+---
+- support exporting investment data/transactions to CSV (#382, #384)
+- more robust sign-in / email handling (#391, #377, #396)
+- fix command-line checking of file extensions (#386)
+
+1.61
+---
+- fix regression: UnboundLocalError for `attention_msg` (#380)
+
+1.60
+---
+- improve handling for new login OTP/soft token selectors (#367, #370)
+- use new categories endpoint, add support for pulling categories from CLI (#359)
+- better long description for pypi
+- refactor sign-in code to its own file (#366)
+- update pandas to latest version
+
 1.59
 ---
 - moved to newest Mint investments endpoint (#342)